Lektion2 Übung zeit

+++ src/lektion2/zeit.java
@@ -13,19 +13,19 @@ public class zeit {
         System.out.println("Die Uhrzeit ist " + stunden + ":" + minuten + ":" + sekunden);
 
         // Sekunden seit Mitternacht
-        int sekSeitMitternacht;
+        double sekSeitMitternacht;
         sekSeitMitternacht = ((stunden * 60) + minuten) * 60 + sekunden;
         System.out.println("Seit Mitternacht sind " + sekSeitMitternacht + " Sekunden vergangen.");
 
         // sekunden bis Mitternacht
-        int sek24H = 24 * 60 * 60;
-        int secBisMitternacht;
+        double sek24H = 24 * 60 * 60;
+        double secBisMitternacht;
         secBisMitternacht = sek24H - sekSeitMitternacht;
         System.out.println("Bis Mitternacht sind es " + secBisMitternacht + " Sekunden.");
 
         // Prozentsatz des vergangen Tages
         double prozentDesTages;
-        prozentDesTages = ((double) sekSeitMitternacht / sek24H) * 100;
+        prozentDesTages = ( sekSeitMitternacht / sek24H) * 100;
         System.out.println("Es sind " + prozentDesTages + " % des Tages vergangen");
